+++ gtatool/src/cmds/component-compute.cpp
@@ -39,6 +39,7 @@
 #include "opt.h"
 #include "cio.h"
 #include "str.h"
+#include "checked_cast.h"
 
 #include "lib.h"
 
@@ -164,7 +165,7 @@ extern "C" int gtatool_component_compute(int argc, char *argv[])
                 hdro.set_compression(gta::none);
                 hdro.write_to(stdout);
                 // Manipulate the GTA data
-                blob element(hdri.element_size());
+                blob element(checked_cast<size_t>(hdri.element_size()));
                 gta::io_state si, so;
                 for (uintmax_t e = 0; e < hdro.elements(); e++)
                 {
