+++ gtatool/src/cmds/dimension-reverse.cpp
@@ -33,6 +33,7 @@
 #include "opt.h"
 #include "cio.h"
 #include "str.h"
+#include "checked_cast.h"
 
 #include "lib.h"
 
@@ -127,7 +128,7 @@ extern "C" int gtatool_dimension_reverse(int argc, char *argv[])
                 hdro.set_compression(gta::none);
                 hdro.write_to(stdout);
                 // Manipulate the GTA data
-                blob element(hdri.element_size());
+                blob element(checked_cast<size_t>(hdri.element_size()));
                 std::vector<uintmax_t> ind(hdro.dimensions());
                 gta::io_state so;
                 for (uintmax_t e = 0; e < hdro.elements(); e++)
