+++ gtatool/src/cmds/merge.cpp
@@ -33,6 +33,7 @@
 #include "opt.h"
 #include "cio.h"
 #include "str.h"
+#include "checked_cast.h"
 
 #include "lib.h"
 
@@ -174,7 +175,7 @@ extern "C" int gtatool_merge(int argc, char *argv[])
                 hdro.component_taglist(c) = hdri[0].component_taglist(c);
             }
             hdro.write_to(stdout);
-            blob element_buf(hdro.element_size());
+            blob element_buf(checked_cast<size_t>(hdro.element_size()));
             std::vector<uintmax_t> indices(hdro.dimensions());
             std::vector<gta::io_state> si(arguments.size());
             gta::io_state so;
